Remove data validations from new IRR columns
When inserting new columns, if the preceding column has data validations,
they get copied to the new cells.
Since the cells we insert are (usually) for IRR computation, the validation is
inapplicable and therefore distracting.
This commit makes it optional, though keeping the behavior for
the column created for conflict resolution, since it's actually useful there.